+++ tests/unit/add1.c
@@ -1,5 +1,0 @@
-void init() {return;}
-
-int add_one(int n) {
-    return n+1;
-}